requirements.txt
Flask==2.1.2
GitPython==3.1.30
mysql-connector-python==8.0.29
bcrypt==4.0.1